Move the internal OpenPGP implementation to its own file.
Split the internal OpenPGP implementation into the bits that are
needed by a new OpenPGP backend like Sequoia, and the bits that are
not needed by another OpenPGP backend.

Move most of the functionality in rpmio/rpmpgp.c into
rpmio/rpmpgp_internal.c.

Leave pgpValStr, and pgpIdentItem, which are used for printing and
needn't be reimplemented by other backends, and pgpReadPkts, which is
just a thin wrapper around pgpParsePkts, and which uses an internal
rpm function that a new backend shouldn't have to worry about
emulating or even calling.

Move the symbol tables, which are used by pgpValStr, pgpIdentItem, and
the internal OpenPGP implementation to rpmio/rpmpgp.h.  These are
common to all implementations.

Fixes #2000.

(backported from d8bb57eeabe249c2c85bf46b1162d7e57a310e37 with
87c4eee816a4ddd4291cf1211c0290f3a731be73 reverted)
